# api/api.yaml
openapi: 3.0.0
servers:
  - url: /oc3
info:
  title: opensvc collector api
  version: 1.0.11
paths:
...
  /nodes/{node_id}/compliance/rulesets/{rset_id}:
    delete:
      description: |
        Detach a ruleset from a node
      operationId: DeleteNodeComplianceRuleset
      parameters:
        - name: node_id
          required: true
          description: The ID of the node from which to detach the ruleset
          in: path
          schema:
            type: string
        - name: rset_id
          required: true
          description: The ID of the ruleset to detach
          in: path
          schema:
            type: string
      responses:
        204:
          description: Ruleset successfully detached
        404:
          description: Node or ruleset not found
        500:
          description: Internal server error
      security:
        - basicAuth: [ ]
        - bearerAuth: [ ]
      tags:
        - collector
    post:
      description: |
        Attach a ruleset to a node
      operationId: PostNodeComplianceRuleset
      parameters:
        - name: node_id
          in: path
          required: true
          description: Node ID
          schema:
            type: string
        - name: rset_id
          in: path
          required: true
          description: Ruleset ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        202:
          description: Ruleset attached
        400:
          $ref: '#/components/responses/400'
        401:
          $ref: '#/components/responses/401'
        403:
          $ref: '#/components/responses/403'
        500:
          $ref: '#/components/responses/500'
      security:
        - basicAuth: [ ]
        - bearerAuth: [ ]
      tags:
        - collector
  /nodes/{node_id}/compliance/modulesets/{mset_id}:
    delete:
      description: |
        Detach a moduleset from a node
      operationId: DeleteNodeComplianceModuleset
      parameters:
        - name: node_id
          required: true
          description: The ID of the node from which to detach the moduleset
          in: path
          schema:
            type: string
        - name: mset_id
          required: true
          description: The ID of the moduleset to detach
          in: path
          schema:
            type: string
      responses:
        204:
          description: Moduleset successfully detached
        404:
          description: Node or moduleset not found
        500:
          description: Internal server error
      security:
        - basicAuth: [ ]
        - bearerAuth: [ ]
      tags:
        - collector
    post:
      description: |
        Attach a moduleset to a node
      operationId: PostNodeComplianceModuleset
      parameters:
        - name: node_id
          in: path
          required: true
          description: Node ID
          schema:
            type: string
        - name: mset_id
          in: path
          required: true
          description: Moduleset ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        202:
          description: Moduleset attached
        400:
          $ref: '#/components/responses/400'
        401:
          $ref: '#/components/responses/401'
        403:
          $ref: '#/components/responses/403'
        500:
          $ref: '#/components/responses/500'
      security:
        - basicAuth: [ ]
        - bearerAuth: [ ]
      tags:
        - collector
...
components:
  schemas:
...
    Problem:
      type: object
      properties:
        detail:
          description: |
            A human-readable explanation specific to this occurrence of the
            problem.
          type: string
        status:
          type: integer
          description: |
            The HTTP status code ([RFC7231], Section 6) generated by the
            origin server for this occurrence of the problem.
        title:
          type: string
          description: |
            A short, human-readable summary of the problem type.  It SHOULD
            NOT change from occurrence to occurrence of the problem, except
            for purposes of localization (e.g., using proactive content
            negotiation; see [RFC7231], Section 3.4).
      required:
        - detail
        - status
        - title
...
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
...
    '400':
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Problem"
    '401':
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Problem"
    '403':
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Problem"
...
    '500':
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Problem"
...